Add svg export to console log
This commit adds a rudimentary function to export the charts to svg. It
is not made user friendly hence it is on this seperate branch.

The core idea is that the options passed to chartjs get passed to a
context which generate a vector graphic. This is then just logged to
the browser console.